Ничего не понятно, но работает
