+++ src/ExceptionBadRequest.java
@@ -1,0 +1,6 @@
+public class ExceptionBadRequest extends RuntimeException{
+    public ExceptionBadRequest() {
+        super("400 Bad Request");
+    }
+
+}
